+++ src/purchases/models.rs
@@ -1,5 +1,5 @@
 use chrono::NaiveDate;
-use diesel::sql_types::{Float, Integer, Nullable, Text};
+use diesel::sql_types::{BigInt, Double, Integer, Nullable, Text};
 use diesel::QueryableByName;
 use serde::Serialize;
 use typescript_definitions::TypeScriptify;
@@ -14,6 +14,7 @@ pub struct RecentPurchase {
     pub vintage: Option<i32>,
     pub memo: Option<String>,
     pub store: Option<String>,
+    #[ts(ts_type = "string | null")]
     pub date: Option<NaiveDate>,
     pub wine_id: i32,
     pub wine_name: Option<String>,
@@ -30,18 +31,18 @@ pub struct RecentPurchase {
 pub struct YearsPurchases {
     #[sql_type = "Integer"]
     pub year: i32,
-    #[sql_type = "Integer"]
-    pub quantity: i32,
-    #[sql_type = "Nullable<Float>"]
-    pub total_price: Option<f32>,
-    #[sql_type = "Nullable<Float>"]
-    pub avg_price: Option<f32>,
+    #[sql_type = "BigInt"]
+    pub quantity: i64,
+    #[sql_type = "Nullable<Double>"]
+    pub total_price: Option<f64>,
+    #[sql_type = "Nullable<Double>"]
+    pub avg_price: Option<f64>,
 }
 
 #[derive(Serialize, TypeScriptify, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct TotalLiters {
-    pub total_liters: f32,
+    pub total_liters: f64,
 }
 
 #[derive(Serialize, TypeScriptify, Debug)]
